+++ packages/brick_sqlite_abstract/test/db/schema_column_test.dart
@@ -1,5 +1,5 @@
 import 'package:test/test.dart';
-import '../../lib/db.dart';
+import 'package:brick_sqlite_abstract/db.dart';
 
 void main() {
   group('SchemaColumn', () {
@@ -23,32 +23,34 @@ void main() {
     group('#forGenerator', () {
       test('simple', () {
         final column = SchemaColumn('first_name', String);
-        expect(column.forGenerator, "SchemaColumn('first_name', String)");
+        expect(column.forGenerator, "SchemaColumn('first_name', columnType: Column.varchar)");
       });
 
       test('primary key', () {
         final column = SchemaColumn('_brick_id', int, autoincrement: true, isPrimaryKey: true);
         expect(
           column.forGenerator,
-          "SchemaColumn('_brick_id', int, autoincrement: true, isPrimaryKey: true)",
+          "SchemaColumn('_brick_id', columnType: Column.integer, autoincrement: true, isPrimaryKey: true)",
         );
       });
 
       test('defaultValue', () {
         final column = SchemaColumn('amount', int, defaultValue: 0);
-        expect(column.forGenerator, "SchemaColumn('amount', int, defaultValue: 0)");
+        expect(column.forGenerator,
+            "SchemaColumn('amount', columnType: Column.integer, defaultValue: 0)");
       });
 
       test('nullable', () {
         final column = SchemaColumn('last_name', String, nullable: false);
-        expect(column.forGenerator, "SchemaColumn('last_name', String, nullable: false)");
+        expect(column.forGenerator,
+            "SchemaColumn('last_name', columnType: Column.varchar, nullable: false)");
       });
 
       test('association', () {
         final column = SchemaColumn('hat_id', int, isForeignKey: true, foreignTableName: 'hat');
         expect(
           column.forGenerator,
-          "SchemaColumn('hat_id', int, isForeignKey: true, foreignTableName: 'hat')",
+          "SchemaColumn('hat_id', columnType: Column.integer, isForeignKey: true, foreignTableName: 'hat', onDeleteCascade: false, onDeleteSetDefault: false)",
         );
       });
 
@@ -56,7 +58,7 @@ void main() {
         final column = SchemaColumn('image', null, columnType: Column.blob);
         expect(
           column.forGenerator,
-          "SchemaColumn('image', null, columnType: Column.blob)",
+          "SchemaColumn('image', columnType: Column.blob)",
         );
       });
     });